test_battle_service/approach_b_agents/requirements.txt
fastapi>=0.128.0
uvicorn[standard]>=0.35.0
sqlalchemy[asyncio]>=2.0.30
aiosqlite>=0.20.0
pydantic>=2.11.0
pydantic-settings>=2.7.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.2.1
email-validator>=2.1.0
